+++ README.md
@@ -1,5 +1,6 @@
-### Working with the virtual environment
+### Useful commands:
 - Creating the environment: `make local-setup`
 - Entering the environment: `poetry shell`
 - Leaving the environment: `exit`
 - Installing libraries in the environment: `make install`
+- Run the Flask application: `python imaginate_api/app.py`
