+++ tests/unit/test_app.py
@@ -6,10 +6,11 @@ import pytest
 import gridfs
 
 # Imaginate modules
-from imaginate_api.app import str_to_bool, validate_id, search_id, build_result, app
+from imaginate_api.utils import str_to_bool, validate_id, search_id, build_result
+from imaginate_api import app
 
 # Other
-from werkzeug.exceptions import BadRequest, HTTPException
+from werkzeug.exceptions import BadRequest, NotFound, HTTPException
 from werkzeug.datastructures import FileStorage
 from bson.objectid import ObjectId
 from io import BytesIO
@@ -45,7 +46,7 @@ def mock_mongo_client():
 
 @pytest.fixture
 def mock_db(mock_mongo_client):
-    mock_db = mock_mongo_client["imaginate"]
+    mock_db = mock_mongo_client["imaginate_dev"]
     return mock_db
 
 
@@ -76,8 +77,8 @@ def mock_data():
 @pytest.fixture(autouse=True)
 def setup(mock_db, mock_fs):
     with (
-        patch("imaginate_api.app.db", mock_db),
-        patch("imaginate_api.app.fs", mock_fs)
+        patch("imaginate_api.db", mock_db),
+        patch("imaginate_api.fs", mock_fs)
         ):
         yield
 
@@ -114,9 +115,9 @@ def test_search_id_success(mock_fs, mock_data):
 
 
 @pytest.mark.parametrize("_id, expected", [
-    ("621f1d71aec9313aa2b9074cd", BadRequest),
-    ("", BadRequest),
-    (None, BadRequest)
+    ("621f1d71aec9313aa2b9074cd", NotFound),
+    ("", NotFound),
+    (None, NotFound)
 ])
 def test_search_id_exception(_id, expected):
     with pytest.raises(HTTPException) as err:
@@ -146,14 +147,14 @@ def test_post_image_create_endpoint_success(client, mock_data):
         assert res.json == build_result(res.json["url"].split("/")[-1], entry["real"], entry["date"], entry["theme"], entry["status"])
 
 @pytest.mark.parametrize("data, expected", [
-    ({}, HTTPStatus.BAD_REQUEST),
+    ({}, HTTPStatus.NOT_FOUND),
     ({
         "date": 0, 
         "theme": "sample", 
         "real": True, 
         "status": "unverified", 
         "file": FileStorage(stream=BytesIO(b"data"), filename="test.pdf", content_type="application/pdf")
-        }, HTTPStatus.BAD_REQUEST)
+        }, HTTPStatus.NOT_FOUND)
 ])
 def test_post_image_create_endpoint_exception(data, expected, client):
     res = client.post("/image/create", data=data, content_type="multipart/form-data")
@@ -199,7 +200,7 @@ def test_get_date_images_endpoint_exception(day, expected, client):
 
 # Tested differently since the endpoint involves sorting
 def test_get_date_latest_endpoint_success(mock_data, client):
-    with patch("imaginate_api.app.fs.find") as mock_find:
+    with patch("imaginate_api.fs.find") as mock_find:
         sorted_data = sorted(mock_data, key=lambda x: x["date"], reverse=True)
         data = iter([Struct(**entry) for entry in sorted_data])
         mock_find.return_value.sort.return_value.limit.return_value = data
@@ -210,7 +211,7 @@ def test_get_date_latest_endpoint_success(mock_data, client):
 
 # Tested differently since the endpoint involves sorting
 def test_get_date_latest_endpoint_exception(client):
-    with patch("imaginate_api.app.fs.find") as mock_find:
+    with patch("imaginate_api.fs.find") as mock_find:
         data = iter([])
         mock_find.return_value.sort.return_value.limit.return_value = data
         res = client.get("date/latest")
